move upload command from FileSystem package to Workspace package
