+++ packages/navigator/src/browser/navigator-contribution.ts
@@ -31,7 +31,6 @@ import { FileNavigatorFilter } from './navigator-filter';
 import { WorkspaceNode } from './navigator-tree';
 import { NavigatorContextKeyService } from './navigator-context-key-service';
 import { TabBarToolbarContribution, TabBarToolbarRegistry, TabBarToolbarItem } from '@theia/core/lib/browser/shell/tab-bar-toolbar';
-import { FileSystemCommands } from '@theia/filesystem/lib/browser/filesystem-frontend-contribution';
 import { NavigatorDiff, NavigatorDiffCommands } from './navigator-diff';
 import { UriSelection } from '@theia/core/lib/common/selection';
 import { PreferenceService } from '@theia/core/lib/browser';
@@ -296,7 +295,7 @@ export class FileNavigatorContribution extends AbstractViewContribution<FileNavi
 
         const downloadUploadMenu = [...NAVIGATOR_CONTEXT_MENU, '6_downloadupload'];
         registry.registerMenuAction(downloadUploadMenu, {
-            commandId: FileSystemCommands.UPLOAD.id,
+            commandId: WorkspaceCommands.UPLOAD.id,
             order: 'a'
         });
         registry.registerMenuAction(downloadUploadMenu, {
